Desktop: MacOS: Add help link if missing encryption password
